Moved the fromJSON code to its own external parser
